token-pipeline/pyproject.toml
[project]
name = "token-pipeline"
version = "0.1.0"
description = "Token Discovery Pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "sqlalchemy",
    "pydantic",
    "pyyaml",
    "apscheduler>=3.10",
    "loguru",
    "anthropic>=0.40",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "isort",
    "mypy",
    "responses",
]

[project.scripts]
token-pipeline = "src.main:main"

[tool.black]
line-length = 88
target-version = ["py310", "py311"]

[tool.isort]
profile = "black"
line_length = 88
